refactor(order): 投影金额由分改为元，移除转换逻辑
- 将 ES、缓存、列表投影中的金额字段类型由 long 改为 BigDecimal
- 直接存储元金额，装配 DTO 时不再进行分转元换算
- 补充元金额一致性的 ES 测试用例

+++ examples/order-example/src/main/java/io/pragmatic/ddd/example/order/domain/order/projection/OrderCacheProjection.java
@@ -2,6 +2,7 @@ package io.pragmatic.ddd.example.order.domain.order.projection;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -39,11 +40,11 @@ public class OrderCacheProjection implements IOrderProjection {
 
     private LocalDateTime paidAt;
 
-    private long totalAmount;
+    private BigDecimal totalAmount;
 
-    private long platformDiscount;
+    private BigDecimal platformDiscount;
 
-    private long actualAmount;
+    private BigDecimal actualAmount;
 
     private CustomerProjection customer;
 
@@ -89,8 +90,8 @@ public class OrderCacheProjection implements IOrderProjection {
         private Long productId;
         private String productName;
         private String spec;
-        private long price;
+        private BigDecimal price;
         private int quantity;
-        private long subtotal;
+        private BigDecimal subtotal;
     }
 }
